src/services/retriever_engine.py
import os
import uuid
from typing import List, Dict, Any, Optional

from pinecone import Pinecone
from langchain.vectorstores import PineconeVectorStore

from src.services.embedding_cache import EmbeddingCache


class RetrieverEngine:
    """
    Retriever engine using Pinecone vector database and cached embeddings.
    """
    # ...existing code... (same as RecommendationEngine but with class name changed)
